+++ packages/antd-management-fast-design-react/src/components/designable/react/widgets/IconWidget/index.jsx
@@ -100,8 +100,8 @@ export const IconWidget = observer((properties) => {
         React.isValidElement(tooltip) || isStr(tooltip)
           ? {}
           : isObj(tooltip)
-          ? tooltip
-          : {};
+            ? tooltip
+            : {};
 
       return (
         <Tooltip {...properties_} title={title}>