+++ packages/example-antd-management-fast-framework/src/modelBuilders/formDesign.js
@@ -6,7 +6,7 @@ import {
   reducerNameCollection,
 } from 'easy-soft-utility';
 
-import { getData, setData } from '../services/formDesign';
+import { getData, setData, setDataSchemaData } from '../services/formDesign';
 
 export function buildModel() {
   return {
@@ -53,6 +53,32 @@ export function buildModel() {
         { call, put },
       ) {
         const response = yield call(setData, payload);
+
+        const dataAdjust = pretreatmentRemoteSingleData({
+          source: response,
+          successCallback: pretreatmentSuccessCallback || null,
+          failCallback: pretreatmentFailCallback || null,
+        });
+
+        yield put({
+          type: reducerNameCollection.reducerRemoteData,
+          payload: dataAdjust,
+          alias,
+          ...reducerDefaultParameters,
+        });
+
+        return dataAdjust;
+      },
+      *setDataSchema(
+        {
+          payload,
+          alias,
+          pretreatmentSuccessCallback,
+          pretreatmentFailCallback,
+        },
+        { call, put },
+      ) {
+        const response = yield call(setDataSchemaData, payload);
 
         const dataAdjust = pretreatmentRemoteSingleData({
           source: response,