+++ packages/antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 2.11.81
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.6.65
+  - antd-management-fast-common@2.6.43
+
 ## 2.11.80
 
 ### Patch Changes